Custom Transitions
I love this slideshow, but it would be a lot better if it had custom transitions. So, here's my fork. I just added a new setting called animation in which you can define the animation you want to use as a string. Then, i used conditional statements to create different values for the visible and hidden transitions based on the value you specify for settings.animation. I created two new transitions.